+++ packages/http/src/validator/index.ts
@@ -13,7 +13,7 @@ import * as Either from 'fp-ts/lib/Either';
 import * as typeIs from 'type-is';
 import { pipe } from 'fp-ts/lib/pipeable';
 import { inRange } from 'lodash';
-import { validateSecurity } from './validators/security/security';
+import { validateSecurity } from './validators/security';
 // @ts-ignore
 import { URI } from 'uri-template-lite';
