+++ packages/http/src/validator/validators/security/handlers/utils.ts
@@ -1,7 +1,7 @@
 import { DiagnosticSeverity } from '@stoplight/types';
 import { Either, left, right } from 'fp-ts/lib/Either';
 import { IPrismDiagnostic } from '@stoplight/prism-core';
-import { IHttpRequest } from '../../../../../types';
+import { IHttpRequest } from '../../../../types';
 
 export type ValidateSecurityFn = (
   someInput: Pick<IHttpRequest, 'headers' | 'url'>,
